+++ src/main/java/duke/Duke.java
@@ -1,6 +1,10 @@
+package duke;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
+import duke.exceptions.UnrecognizedException;
+import duke.exceptions.MissingInputException;
 
 public class Duke {
     public static void main(String[] args) throws FileNotFoundException, IOException {
