+++ src/main/java/duke/Storage.java
@@ -1,10 +1,15 @@
+package duke;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.util.Scanner;
 import java.io.FileWriter;
-
+import duke.TaskList;
+import duke.tasks.Task;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Todo;
 
 public class Storage {
     // Parses file and outputs it as a list
