+++ src/main/java/duke/data/saved_tasks.txt
@@ -1,0 +1,3 @@
+todo true Read
+deadline true 1999-12-03 Return book 
+deadline false 2022-03-03 Eat book
