+++ src/main/java/duke/exceptions/MissingInputException.java
@@ -1,3 +1,4 @@
+package duke.exceptions;
 public class MissingInputException extends Exception {
     public MissingInputException(String message) {
         super(message);
