+++ src/main/java/duke/exceptions/UnrecognizedException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class UnrecognizedException extends Exception {
     public UnrecognizedException(String message) {
         super(message);
